Cambios del skybox y estructura sus archivos
